redis-langcache-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1
sentence-transformers==4.0.1
redis==5.0.1
numpy==1.26.4
python-dotenv==1.0.0
transformers==4.50.3
torch==2.6.0
